--- backend-api/index.js
@@ -1,41 +1,0 @@
-const express = require("express");
-const bodyParser = require("body-parser");
-const cors = require("cors");
-const helmet = require("helmet");
-const morgan = require("morgan");
-const mongoose = require("mongoose");
-
-const routes = require("./routes/routes");
-
-const app = express();
-require("dotenv").config();
-const mongoString = process.env.DATABASE_URL;
-
-mongoose.connect(mongoString);
-const database = mongoose.connection;
-
-database.on("error", (error) => {
-  console.log(error);
-});
-
-database.once("connected", () => {
-  console.log("Database Connected");
-});
-
-const posts = [{ title: "Hello, mysore!" }];
-
-app.use(helmet());
-app.use(bodyParser.json());
-app.use(cors());
-app.use(morgan("combined"));
-
-app.use("/api", routes);
-
-app.get("/", (req, res) => {
-  res.send(posts);
-});
-
-const PORT = process.env.PORT || 3001;
-app.listen(PORT, () => {
-  console.log(`Server running on port ${PORT}`);
-});
